03/06 soir, submit questionnaire et calcul note en cours

+++ src/main/java/repository/ApprenantRepository.java
@@ -1,12 +1,11 @@
 package repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
 import model.Apprenant;
 
 
-public interface ApprenantRepository extends CrudRepository<Apprenant, Long>{
+public interface ApprenantRepository extends CrudRepository<Apprenant, Integer>{
